fix: resolve clippy lints for nightly-2026-04-24 and matrix-sdk compat
Fix all new clippy lints introduced by the nightly bump:
- collapsible_match: collapse inner `if` into match arm guards
- unnecessary_sort_by: use sort_by_key with Reverse for desc sorts
- question_mark: use `?` operator for early returns
- useless_conversion: remove redundant .into_iter()
- unnecessary cast: remove u32 -> u32 casts
- explicit_counter_loop: use enumerate() instead of manual counter

Also work around matrix-sdk 0.16 query-depth overflow on nightly
>= 2025-12-28 by removing `matrix` from default features and making
the metrics feature forward conditional (`moltis-matrix?/metrics`).
Matrix support is still available via `--features matrix`.

+++ crates/cron/src/heartbeat.rs
@@ -143,9 +143,9 @@ pub fn is_within_active_hours(start: &str, end: &str, timezone: &str) -> bool {
     let end_minutes = if end == "24:00" {
         24 * 60
     } else {
-        end_time.hour() as u32 * 60 + end_time.minute() as u32
+        end_time.hour() * 60 + end_time.minute()
     };
-    let start_minutes = start_time.hour() as u32 * 60 + start_time.minute() as u32;
+    let start_minutes = start_time.hour() * 60 + start_time.minute();
 
     let now_minutes = current_minutes(timezone);
